requirements.txt
pyyaml
Jinja2
tox
